Revert "Move I/O system to io dir"
