+++ configure.ac
@@ -287,18 +287,18 @@ AC_CONFIG_LINKS([include/souffle/ExplainProvenance.h:src/ExplainProvenance.h])
 AC_CONFIG_LINKS([include/souffle/ExplainProvenanceImpl.h:src/ExplainProvenanceImpl.h])
 AC_CONFIG_LINKS([include/souffle/ExplainTree.h:src/ExplainTree.h])
 AC_CONFIG_LINKS([include/souffle/EquivalenceRelation.h:src/EquivalenceRelation.h])
-AC_CONFIG_LINKS([include/souffle/IOSystem.h:src/io/IOSystem.h])
+AC_CONFIG_LINKS([include/souffle/IOSystem.h:src/IOSystem.h])
 AC_CONFIG_LINKS([include/souffle/LambdaBTree.h:src/LambdaBTree.h])
 AC_CONFIG_LINKS([include/souffle/Logger.h:src/Logger.h])
 AC_CONFIG_LINKS([include/souffle/PiggyList.h:src/PiggyList.h])
 AC_CONFIG_LINKS([include/souffle/ProfileDatabase.h:src/ProfileDatabase.h])
 AC_CONFIG_LINKS([include/souffle/ProfileEvent.h:src/ProfileEvent.h])
 AC_CONFIG_LINKS([include/souffle/RamTypes.h:src/RamTypes.h])
-AC_CONFIG_LINKS([include/souffle/ReadStream.h:src/io/ReadStream.h])
-AC_CONFIG_LINKS([include/souffle/ReadStreamCSV.h:src/io/ReadStreamCSV.h])
-AC_CONFIG_LINKS([include/souffle/ReadStreamSQLite.h:src/io/ReadStreamSQLite.h])
-AC_CONFIG_LINKS([include/souffle/ReadStreamJSON.h:src/io/ReadStreamJSON.h])
-AC_CONFIG_LINKS([include/souffle/SerialisationStream.h:src/io/SerialisationStream.h])
+AC_CONFIG_LINKS([include/souffle/ReadStream.h:src/ReadStream.h])
+AC_CONFIG_LINKS([include/souffle/ReadStreamCSV.h:src/ReadStreamCSV.h])
+AC_CONFIG_LINKS([include/souffle/ReadStreamSQLite.h:src/ReadStreamSQLite.h])
+AC_CONFIG_LINKS([include/souffle/ReadStreamJSON.h:src/ReadStreamJSON.h])
+AC_CONFIG_LINKS([include/souffle/SerialisationStream.h:src/SerialisationStream.h])
 AC_CONFIG_LINKS([include/souffle/SignalHandler.h:src/SignalHandler.h])
 AC_CONFIG_LINKS([include/souffle/SouffleInterface.h:src/SouffleInterface.h])
 AC_CONFIG_LINKS([include/souffle/SymbolTable.h:src/SymbolTable.h])
@@ -316,10 +316,10 @@ AC_CONFIG_LINKS([include/souffle/utility/StreamUtil.h:src/utility/StreamUtil.h])
 AC_CONFIG_LINKS([include/souffle/utility/StringUtil.h:src/utility/StringUtil.h])
 AC_CONFIG_LINKS([include/souffle/utility/ParallelUtil.h:src/utility/ParallelUtil.h])
 AC_CONFIG_LINKS([include/souffle/utility/tinyformat.h:src/utility/tinyformat.h])
-AC_CONFIG_LINKS([include/souffle/WriteStream.h:src/io/WriteStream.h])
-AC_CONFIG_LINKS([include/souffle/WriteStreamCSV.h:src/io/WriteStreamCSV.h])
-AC_CONFIG_LINKS([include/souffle/WriteStreamSQLite.h:src/io/WriteStreamSQLite.h])
-AC_CONFIG_LINKS([include/souffle/WriteStreamJSON.h:src/io/WriteStreamJSON.h])
+AC_CONFIG_LINKS([include/souffle/WriteStream.h:src/WriteStream.h])
+AC_CONFIG_LINKS([include/souffle/WriteStreamCSV.h:src/WriteStreamCSV.h])
+AC_CONFIG_LINKS([include/souffle/WriteStreamSQLite.h:src/WriteStreamSQLite.h])
+AC_CONFIG_LINKS([include/souffle/WriteStreamJSON.h:src/WriteStreamJSON.h])
 AC_CONFIG_LINKS([include/souffle/gzfstream.h:src/gzfstream.h])
 AC_CONFIG_LINKS([include/souffle/json11.h:src/json11.h])
 AC_CONFIG_LINKS([include/souffle/profile/Cell.h:src/profile/Cell.h])
