+++ src/WriteStream.h
