+++ src/WriteStreamSQLite.h
