+++ src/interpreter/InterpreterEngine.cpp
@@ -17,22 +17,23 @@
 #include "AggregateOp.h"
 #include "BinaryConstraintOps.h"
 #include "FunctorOps.h"
+#include "Global.h"
+#include "IOSystem.h"
 #include "Logger.h"
 #include "ProfileEvent.h"
 #include "RamTypes.h"
+#include "ReadStream.h"
 #include "RecordTable.h"
 #include "SignalHandler.h"
 #include "SymbolTable.h"
 #include "TypeAttribute.h"
+#include "WriteStream.h"
 #include "interpreter/InterpreterContext.h"
 #include "interpreter/InterpreterGenerator.h"
 #include "interpreter/InterpreterIndex.h"
 #include "interpreter/InterpreterNode.h"
 #include "interpreter/InterpreterPreamble.h"
 #include "interpreter/InterpreterRelation.h"
-#include "io/IOSystem.h"
-#include "io/ReadStream.h"
-#include "io/WriteStream.h"
 #include "ram/Aggregate.h"
 #include "ram/AutoIncrement.h"
 #include "ram/Break.h"
